mqtt_to_kafka/requirements.txt

paho-mqtt
kafka-python